requirements.txt
PyYAML==6.0.1
alembic==1.13.2
SQLAlchemy==2.0.31
python-magic==0.4.27
pillow==10.4.0
img2pdf==0.5.1
docx2pdf==0.1.8
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
shiboken6==6.7.1